Remove proxyQueryExecutionContext
